src/tools/marker_mcp/requirements.txt
mcp
marker-pdf
